backend/app/core/security.py
import os
import jwt
from datetime import datetime, timedelta, timezone
from passlib.context import CryptContext

# 配置 bcrypt 密码加密器
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# 💡 直接从环境变量读取，找不到时给个默认值防崩
SECRET_KEY = os.getenv("SECRET_KEY", "fallback_secret_key")
ALGORITHM = os.getenv("ALGORITHM", "HS256")
ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 10080))

def get_password_hash(password: str) -> str:
    """将明文密码加密成哈希值"""
    return pwd_context.hash(password)

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """校验密码是否正确"""
    return pwd_context.verify(plain_password, hashed_password)

def create_access_token(data: dict) -> str:
    """生成 JWT Token"""
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt