backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
asyncpg==0.29.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
python-multipart>=0.0.6
Pillow>=10.1.0